+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Confirm button is not disabled after the click and each click on this button sends the request to the server. Refs UIU-2645.
 * When formatting names, correctly treat empty-strings as missing. Refs UIU-2610.
 * Additional updates to Pop-up note Refs UIU-2653
+* Modify error message when a Fee/Fine Owner enters a Fee/Fine Type that is duplicated by "Shared" Fee/Fine Owner. Refs UIU-2560.
 
 ## [8.1.0](https://github.com/folio-org/ui-users/tree/v8.1.0) (2022-06-27)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.0.0...v8.1.0)